Eliminated more invalid pointers with nonzero base. READ_BYTE/LINEAR2UNIX don't work well with linear DPMI addresses yet!
git-svn-id: https://dosemu.svn.sourceforge.net/svnroot/dosemu/trunk@1929 7758b857-130e-0410-ab9c-d8735a528d18

+++ src/arch/linux/debugger/mhpdbgc.c
@@ -1430,7 +1430,7 @@ int mhp_bpchk(unsigned int a1)
 int mhp_getcsip_value()
 {
   int  seg, off, limit;
-  if (IN_DPMI) return (int)(uintptr_t)mhp_getadr("cs:eip", &seg, &off, &limit);
+  if (IN_DPMI) return mhp_getadr("cs:eip", &seg, &off, &limit);
   else return (LWORD(cs) << 4) + LWORD(eip);
 }
